[ISSUE #2980]♻️Refactor ConsumeQueueStore💫
