+++ rocketmq-store/src/base/message_store.rs
@@ -48,8 +48,8 @@ use crate::hook::put_message_hook::PutMessageHook;
 use crate::hook::send_message_back_hook::SendMessageBackHook;
 use crate::log_file::commit_log::CommitLog;
 use crate::log_file::mapped_file::MappedFile;
+use crate::queue::consume_queue_store::ConsumeQueueStoreTrait;
 use crate::queue::ArcConsumeQueue;
-use crate::queue::ConsumeQueueStoreTrait;
 use crate::stats::broker_stats_manager::BrokerStatsManager;
 use crate::store::running_flags::RunningFlags;
 use crate::store_error::StoreError;
@@ -473,7 +473,7 @@ pub trait MessageStoreInner: Sync + 'static {
     //fn get_perf_counter(&self) -> Arc<PerfCounterTicks>;
 
     /// Get the queue store
-    fn get_queue_store(&self) -> &dyn ConsumeQueueStoreTrait;
+    fn get_queue_store(&self) -> &Box<dyn ConsumeQueueStoreTrait>;
 
     /// If 'sync disk flush' is configured in this message store
     fn is_sync_disk_flush(&self) -> bool;
