+++ rocketmq-store/src/queue/batch_consume_queue.rs
@@ -32,8 +32,9 @@ use crate::config::message_store_config::MessageStoreConfig;
 use crate::consume_queue::mapped_file_queue::MappedFileQueue;
 use crate::filter::MessageFilter;
 use crate::log_file::mapped_file::default_mapped_file_impl::DefaultMappedFile;
+use crate::queue::consume_queue::ConsumeQueueTrait;
 use crate::queue::queue_offset_operator::QueueOffsetOperator;
-use crate::queue::ConsumeQueueTrait;
+use crate::queue::referred_iterator::ReferredIterator;
 use crate::queue::CqUnit;
 use crate::queue::FileQueueLifeCycle;
 
@@ -236,12 +237,12 @@ impl ConsumeQueueTrait for BatchConsumeQueue {
     }
 
     #[inline]
-    fn get_earliest_unit(&self) -> CqUnit {
+    fn get_earliest_unit(&self) -> Option<CqUnit> {
         todo!()
     }
 
     #[inline]
-    fn get_latest_unit(&self) -> CqUnit {
+    fn get_latest_unit(&self) -> Option<CqUnit> {
         todo!()
     }
 
@@ -267,15 +268,6 @@ impl ConsumeQueueTrait for BatchConsumeQueue {
 
     #[inline]
     fn get_offset_in_queue_by_time(&self, timestamp: i64) -> i64 {
-        todo!()
-    }
-
-    #[inline]
-    fn get_offset_in_queue_by_time_boundary(
-        &self,
-        timestamp: i64,
-        boundary_type: BoundaryType,
-    ) -> i64 {
         todo!()
     }
 
@@ -339,16 +331,23 @@ impl ConsumeQueueTrait for BatchConsumeQueue {
     }
 
     #[inline]
-    fn iterate_from(&self, start_index: i64) -> Option<Box<dyn Iterator<Item = CqUnit> + Send>> {
+    fn iterate_from(&self, start_index: i64) -> Option<Box<dyn ReferredIterator<CqUnit>>> {
         todo!()
     }
 
-    #[inline]
-    fn iterate_from_inner(
+    fn iterate_from_with_count(
         &self,
         start_index: i64,
         count: i32,
-    ) -> Option<Box<dyn Iterator<Item = CqUnit> + Send>> {
+    ) -> Option<Box<dyn ReferredIterator<CqUnit>>> {
+        todo!()
+    }
+
+    fn get_offset_in_queue_by_time_with_boundary(
+        &self,
+        timestamp: i64,
+        boundary_type: BoundaryType,
+    ) -> i64 {
         todo!()
     }
 }
