+++ rocketmq-store/src/queue/build_consume_queue.rs
@@ -18,10 +18,9 @@ use rocketmq_common::common::sys_flag::message_sys_flag::MessageSysFlag;
 
 use crate::base::commit_log_dispatcher::CommitLogDispatcher;
 use crate::base::dispatch_request::DispatchRequest;
+use crate::queue::consume_queue_store::ConsumeQueueStoreInterface;
 use crate::queue::local_file_consume_queue_store::ConsumeQueueStore;
-use crate::queue::ConsumeQueueStoreTrait;
 
-#[derive(Clone)]
 pub struct CommitLogDispatcherBuildConsumeQueue {
     consume_queue_store: ConsumeQueueStore,
 }
